# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Corridor API
  description: |
    Corridor is a Business Orchestration Platform (BOAT) providing payment infrastructure,
    social payments, crowdfunding goals, and Earned Wage Access (EWA) services.
    
    ## Authentication
    Most endpoints require authentication via Bearer token or API key:
    - **Bearer Token**: For user-facing applications (login required)
    - **API Key**: For partner integrations (header: `X-API-Key`)
    
    ## Rate Limits
    - Auth endpoints: 10 requests/minute
    - Payment endpoints: 100 requests/minute  
    - Partner API: 1000 requests/minute (with API key)
  version: 1.0.0
  contact:
    email: people@corridormoney.net
    name: Corridor Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.corridormoney.net
    description: Production
  - url: https://api-sandbox.corridormoney.net
    description: Sandbox
  - url: http://localhost:8000
    description: Local Development
...
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
        '409':
          description: User already exists
...
  /api/auth/verify/send:
    post:
      tags: [Authentication]
      summary: Send verification code
      operationId: sendVerificationCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, contact]
              properties:
                channel:
                  type: string
                  enum: [email, whatsapp]
                contact:
                  type: string
                  description: Email address or phone number
      responses:
        '200':
          description: Code sent
        '429':
          description: Rate limit exceeded
  /api/auth/verify/confirm:
    post:
      tags: [Authentication]
      summary: Verify code
      operationId: verifyCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contact, code]
              properties:
                contact:
                  type: string
                code:
                  type: string
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
...
  /api/social/pay:
    post:
      tags: [Social Payments]
      summary: Send a social payment
      operationId: createSocialPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialPaymentRequest'
      responses:
        '200':
          description: Payment sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
  /api/wallets:
    get:
      tags: [Wallets]
      summary: List user's wallets
      operationId: getWallets
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wallet'
    post:
      tags: [Wallets]
      summary: Create a new wallet
      operationId: createWallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currency]
              properties:
                currency:
                  type: string
                  enum: [USDC, USD, KES, NGN, SOL]
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    RegisterRequest:
      type: object
      required: [email, name, password, type]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
          minLength: 8
        type:
          type: string
          enum: [PERSONAL, BUSINESS]
        phone:
          type: string
        country:
          type: string
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        user:
          $ref: '#/components/schemas/Account'
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        account_type:
          type: string
          enum: [PERSONAL, BUSINESS]
        country:
          type: string
        kyc_status:
          type: string
        onboarding_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sender_wallet_id:
          type: string
          format: uuid
        recipient_wallet_id:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        message:
          type: string
        created_at:
          type: string
          format: date-time
    SocialPaymentRequest:
      type: object
      required: [amount]
      properties:
        from_wallet:
          type: string
          format: uuid
        to_wallet:
          type: string
          format: uuid
        to_email:
          type: string
        to_handle:
          type: string
        amount:
          type: number
        currency:
          type: string
        message:
          type: string
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        currency:
          type: string
        balance:
          type: number
        address:
          type: string
        created_at:
          type: string
          format: date-time
...